neet-test-backend/src/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
google-generativeai==0.3.2
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
requests==2.31.0
urllib3==2.0.5
openai==1.3.0
python-dateutil==2.8.2
gunicorn==20.1.0